Add back old css
I realised I forgot to add the css back to the website.
I will probably re-write the css in a similar fashion because
I believe that it is broken at the moment. It also has a lot
of styles from the original portfolio home page, so I will
get rid of those.

When I rewite it I will try to keep it in the same kind of style
I used then.