Fix bug causing poll options to be unselectable
If there were two options that happened to be the same, it became very
hard to click any one except first one. This was because the id used for
the label was the name of the option. If two options had the same name, it
would select the first one.

To fix this I changed it so it uses the index of the option as the id.